# Conflicts:
#	src/main/java/com/skyblock/skyblock/Skyblock.java

+++ src/main/java/com/skyblock/skyblock/features/enchantment/enchantments/sword/EnderSlayerEnchantment.java
@@ -13,7 +13,7 @@ import java.util.function.Function;
 
 public class EnderSlayerEnchantment extends SwordEnchantment {
 
-    private static final Function<Integer, Double> getDamage = (level) -> Util.createFetchableDictionary(level - 1, 15, 30, 45, 60, 80, 100);
+    private static final Function<Integer, Double> getDamage = (level) -> (double) Util.createFetchableDictionary(level - 1, 15, 30, 45, 60, 80, 100);
 
     public EnderSlayerEnchantment() {
         super("ender_slayer", "Ender Slayer", (level) -> {
